demo/requirements.txt
Flask==2.2.2
google-genai